Move modin test to 3.7 job
