+++ LICENSE
@@ -289,8 +289,8 @@ Code in python/ray/_private/prometheus_exporter.py is adapted from https://githu
 # limitations under the License.
 
 --------------------------------------------------------------------------------
-Code in python/ray/tests/test_modin and python/ray/tests/modin_test_utils adapted 
-from:
+Code in python/ray/tests/modin/test_modin and 
+python/ray/tests/modin/modin_test_utils adapted from:
 - http://github.com/modin-project/modin/master/modin/pandas/test/test_general.py
 - http://github.com/modin-project/modin/master/modin/pandas/test/utils.py
