+++ python/ray/tests/BUILD
@@ -65,7 +65,6 @@ py_test_module_list(
     "test_memory_scheduling.py",
     "test_plasma_unlimited.py",
     "test_metrics.py",
-    "test_modin.py",
     "test_multi_node.py",
     "test_multi_node_2.py",
     "test_multi_node_3.py",
