+++ python/ray/tests/modin/BUILD
@@ -1,0 +1,7 @@
+py_test(
+ name = "test_modin",
+ size = "small",
+ srcs = ["test_modin.py"],
+ deps = ["//:ray_lib"],
+ tags = ["exclusive"],
+)
